shill: Destroy services on modem stop without waiting for MM response
During modem stop operation, if MM does not respond to Dbus calls,
we end up with stale cellular device object in shill as cellular services
are not cleared. This is a change to destroy services without waiting
for a response from MM. We are expected to cleanup services irrespective
result of stop operation.

BUG=b:230401667
TEST=Tested on Nissa with mmcli modem reset command

Change-Id: I3f368160cb1927c09bef3a1f3c665e0ec4e513ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4989586
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Ujjwal Pande <ujjwalpande@google.com>
Tested-by: Ujjwal Pande <ujjwalpande@google.com>