fbpreprocessor: check if the feature is allowed by policy
Only allow firmware dumps to be added to feedback reports if the
UserFeedbackWithLowLevelDebugDataAllowed policy allows it. Query the
value of that policy every time a user logs in.

To avoid issues if 2 users are logged in and they have different policy
settings regarding the inclusion of firmware dumps in feedback reports,
check that we only have 1 active session.

Also add an allowlist of accounts that are allowed to add firmware dumps
to feedback reports.

BUG=b:291346820
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord
TEST=log in guest session, verify in the log that there is no primary
user but there is 1 active session
TEST=log in with a "real" user, verify in the log that the primary user
is detected and that there is 1 active session
TEST=http://gpaste/6142262005202944
TEST=policy retrieval code not tested yet, pending availability of the
policy. fbpreprocessord is not installed on shipping boards anyway so it
should be ok for now.

Change-Id: I5694c03cb9defe55f665a3804be40d0e14cea7b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5008680
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Yichen Yu <yichenyu@google.com>