cryptohome: Use RunWhenAvailable for PrepareAuthSession
This converts the PrepareAuthSession code over to using the
RunWhenAvailable mechanism that is used with terminate. In general this
this approach (enqueuing the request to run when the session is not in
use) is probably a better choice for all operations, not just terminate.

This change just converts one additional function, but it lays a bit of
groundwork to try and reduce the amount of duplication and callback
noise.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I6a71f30be8598fde74338ff0547e0491ee805c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4996221
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>