net-base: polish NDUserOptionHeader struct
The format of the neighbor discovery user option is a TLV structure.
The common part only contains the type field and the length field.
This CL removes the reserved and the lifetime fields from the
NDUserOptionHeader and parse them at the separated following structure.

BUG=b:305129516
TEST=FEATURES=test emerge-brya net-base
TEST=/usr/libexec/fuzzers/rtnl_handler_fuzzer

Change-Id: I22d61af177d12289883f015925145e9bfb1b6eee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4996612
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>