patchpanel: Implement ebpf program for WebRTC detection
Implement the eBPF program for matching if a packet is a UDP DTLS client
hello packet which contains the "use_srtp" extension. This program will
be used by iptables in detection for WebRTC connection.

Also adds the corresponding unit tests for the eBPF program.

BUG=b:296952085
BUG=b:296959393
TEST=FEATURES="test" emerge-$BOARD patchpanel

Change-Id: Ief975d74b13906428a4028347c3e2619c8512c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4951019
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Damien Dejean <damiendejean@google.com>