cryptohome: Use RunWhenAvailable for more UDA operations
This converts both some "easy" functions as well of a couple of more
complex ones over to the new model of queuing up work instead of
executing it directly.

For the "easy" functions we have GetAuthSessionStatus and
GetRecoveryRequest where the change is just hoisting the very small
function bodies into a lambda supplied to our RunWhenAvailable utility.
These are the only remaining uses of FindAuthSession which are trivial
to convert.

This also converts AuthenticateAuthFactor, which is a big change but
which still mostly just amounts to hoisting almost all of the function
into a separate impl helper. And it converts CreatePersistentUser which
is a more invasive change that requires actually reworking the cuts of
the method somewhat, replacing the existing helper function with a new
one. The new code also slightly adjusted in behavior in that it does
only a single lookup (via RunWhenAvailable) instead of two separate
FindAuthSession calls. This is actually an improvement.

There are still several more functions and helpers that use Find in
subsequent commits.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I9ba8bb6a766b1840bfccd8888873051295c1595a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4996860
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>