hiberman: Record hibernating user after login
hiberman currently tries to obtain the account id of the primary
user at hibernate time. This doesn't work because userspace is
frozen (by powerd) when 'hiberman hibernate' is invoked, so
session manager can't respond to the request.

Instead record the 'hibernating user' after login. This is not
entirely correct semantically, since the user might not actually
hibernate, but it works and is less involved than other
solutions.

BUG=b:308631058
TEST=hibernate and resume

Change-Id: I2cf650cb0afb456ea28d0b8f721bd73b715483e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5005362
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>