patchpanel: Install iptables rules for WebRTC detection
This patch adds code to install the iptables rules for WebRTC detection
in Datapath. Since iptables needs the bpf() syscall to load the bpf
program, we also need to update several places for CAP_BPF and the
seccomp filter.

After this patch, the WebRTC detection will be enabled when qos flag is
on.

BUG=b:296952085
TEST=Manually checked that the rules were installed correctly.
TEST=tast run $DUT network.QosWebRTC

Change-Id: I11e2b23b05f1efdfd0248b60e67edb76d0892932
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5007477
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>