cryptohome: Convert two Prepare Vault functions to RunWhen
This converts PreparePersistentVault and PrepareVaultForMigration over
to the RunWhen model. This lets us get rid of the second-last use of
FindAuthSession, just leaving GetAuthenticatedAuthSession as the only
remaining usage (unfortunately still used by a bunch of operations).

This also provides the model for replacing GetAuthenticatedAuthSession
by providing a helper template, similar to the existing one that a lot
of other functions were converted over to, that basically does the same
thing but then also does the "is the session authed" check. So more
followup changes should eventually be able to convert the remaining
operations over to using it.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Idf1171e84fb0da3c151bcc08f03eeead5efb358b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5008811
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>