fbpreprocessor: handle failure to start pseudonymizations
If we fail to post the pseudonymization task we should not count that
attempt against the rate limit since it did not actually start. Only add
the timestamp if the pseudonymization was successfully posted.

Also protect accesses to the list of recent operations with a lock since
potentially different threads could request pseudonymizations.

Also add comments to class members/methods for documentation.

BUG=b:291346795
TEST=cros-workon-${BOARD} start fbpreprocessord
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord

Change-Id: I65d78a47a198b62dfed305cbeda15382217011f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5008815
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Yichen Yu <yichenyu@google.com>