flex_hwis: handle update/delete error responses
ChromeOS Flex Hardware Information Service (flex_hwis)
categorizes the outcomes of update and delete API calls
into three types: success, error, and device not found
on server.

For delete requests, if the device is not found on the
server, the device is considered to have been deleted
successfully. In this case, the device record on the
client can be deleted.

For update requests, if the device is not found on the
server, flex_hwis should register the device on the
server again.

Logs should be added to each request and response to
make it easier to track the status of API calls in
the future.

BUG=b:307742905
TEST=emerge-reven chromeos-base/flex_hwis
TEST=cros deploy vm chromeos-base/flex_hwis
TEST=run flex_hwis_tool service at login
TEST=cros_workon_make --board=reven chromeos-base/flex_hwis --test

Change-Id: I6143a46a509abcb9882fa9f9e32bfa756ccc70ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5008723
Tested-by: Ting-Hao Lin <tinghaolin@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>