shill: create a new transport object on every new entitlement check
We also use the existence of the transport object to flag that there
is a request ongoing, no need to have a separate boolean field for
that.

The unit tests are updated to allow setting a testing-specific
transport object, which will be used instead of creating a new one
during each request.

BUG=b:300161883
TEST=Manual tests in AT&T/brya/vell/FM350 device.
  Run CheckTetheringReadiness() before/after enabling tethering in the
  setup that requires "DUN as DEFAULT" logic.

Change-Id: I3b832c8269b07b4f13817e6fb4958498a2c74c34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5009375
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Madhav <madhavadas@google.com>