cryptohome: Use OnDoneCallback throughout UDA
In a prior change I added the OnDoneCallback alias to make the intent
of some new calls somewhat clearer, but to avoid adding a ton of noise
to that change I did not touch any existing declarations. This change
goes back and changes all of the UserDataAuth functions to use the alias
for all of the on_done parameters.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I1ed74849765cd45a8180889fd6edabad832cbb78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5010643
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>