hiberman: Always hibernate to S5
S4 isn't well supported (if at all) on Chrome OS devices. Intel
Keylocker is the only reason for hibernating to S4. We now block
hibernate in the kernel when Keylocker is in use (which is
supposed to be rare), so let's get rid of S4 support and always
hibernate to S5.

BUG=b:309149035
TEST=hibernate and resume on a vPro device

Cq-Depend: chromium:5010240
Change-Id: Ic926994e39e1dbe55838e30fe2003cb72d47f47e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5011201
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Auto-Submit: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>