cryptohome: Convert more UDA operations over to RunWhenAvailable
This takes care of another batch, mostly the "easy" ones using
GetAuthenticatedAuthSession. These are the operations which are already
async-ready and so we can just replace the GetSession call with the
appropriate RunWith helper, lifting the rest of the function body into
the RunWith operation.

The only remaining uses after this that need to be fixed are
GetHibernateSecret, which needs to be made async (non trivial but fairly
mechanical and straightforward) and StartMigrateToDircrypto (not
entirely obvious, may need some significant rearrangement).

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I207bdc57588ff6457e720c8bd99c6fb810e952b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5015324
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>