hiberman: Don't clobber resume errors if the subsequent hibermeta mount fails
If mounting hibermeta fails in ResumeConductor::resume() after
resume_inner() returned an error the previous code returned the
error of the hibermeta mount and the error returned by
resume_inner() was lost. Change this to logging the hibermeta
mount error and returning the resume_inner() error, so it can
be logged by the caller.

BUG=b:309654982
TEST=manual

Change-Id: Ia76fb60cec365d160e661f177398da67fa1e225d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5014386
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>