metrics: Exempt audio_peripheral project to always be recorded to the disk
Allow audio_peripheral project to always be recorded to the disk
disregarding the UMA consent status. This is due to the known
flakiness where structured metrics may not be properly saved to the
disk despite UMA consent is on. This exemption shall be removed once
the flakiness is fixed.

If UMA consent is off, this bypass would save audio_peripheral
metrics to the disk, but Chromium Metrics Reader would not report
it to Google servers. The saved metrics are removed from the disk
after 10 minutes.

BUG=b:286796175
BUG=b:309565425
TEST=emerge-${BOARD} metrics

Change-Id: I20a0c707d0336a0dbc4659ae56885bae75003a2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5015988
Auto-Submit: Judy Hsiao <judyhsiao@google.com>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Jong Ahn <jongahn@chromium.org>
Tested-by: Judy Hsiao <judyhsiao@google.com>