camera: Fix SensorHalClient reconnection
CancellationRelay shouldn't cancel all futures when iioservice is
disconnected. It should only be used when SensorHalClientImpl is
destructed. The previous implementation was incorrect to assume that
CancellationRelay will be reset after calling `CancelAllFutures()`.

This CL also fixes an issue of RegisterSamplesObserver: it should check
if iioservice is connected. If not, constructing SensorReader (i.e. start
reading samples) should be delayed.

BUG=b:144747001
TEST=run on Quackingstick (strongbad)

Change-Id: Ifd575c34c5c2141b592f16de5c0d609da38024eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5016366
Tested-by: Harvey Yang <chenghaoyang@chromium.org>
Commit-Queue: Harvey Yang <chenghaoyang@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>