cryptohome: Remove the last of FindAuthSession
This converts StartMigrateToDircrypto over to using RunWhenAvailable,
and it being the last usage of FindAuthSession means we can now get rid
of it from UserDataAuth.

This requires a bit of re-arrangement of how the function works, and
similar to GetHibernateSecret it also requires adding error_info to the
reply protobuf. But functionally the new behavior should be fairly
equivalent.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I21bf360e1b5464f7b393d43f2b27a06743d786c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5018597
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>