cryptohome: Get rid of FindAuthSession
Now that nothing uses it, make RunWhenAvailable the ONLY way to do stuff
with a managed AuthSession.

This also allows the InUse object to be simplified somewhat, because we
can now remove the "is session active" bit. The only use of that was the
"you tried to find a session which is valid, but busy" but that's now
not actually possible anymore: either you got the object from Create/Add
in which case it's guaranteed to not be busy, or you got it from
RunWhenAvailable where the object can only be invalid if the session
does not exist.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ida9665b1880bce987ad1a416976c3a85f69d12f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5018608
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>