init: Fix condition to accommodate Flex devices
Currently, we assume that if the display was initialized in firmware
it must be with our logo, so there is no need to call frecon.
However, this assumption is false because on some Flex devices,
the OEMs show their own logo from firmware (using the same
functionality as our fw splash feature).

Fix this issue by checking a cros_config parameter that specifies
whether or not Google's logo was displayed from firmware.

BUG=b:302505216, b:309508964
TEST=Manually checked backlight is set and frecon displays logo
	 on Reven + on Rex without fw_splash feature
	 &&
	 Manually checked backlight is not changed + frecon does not
	 modeset the display when on Rex with fw_splash feature

Cq-Depend: chromium:4973423
Change-Id: I92b945c208dd1726aeb430ed630df0313bc65f67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5019336
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Paz Zcharya <pazz@google.com>
Commit-Queue: Paz Zcharya <pazz@google.com>