resourced: schedqos: restore states from file
Resourced uses a file on tmpfs to store every process/thread states.
Resourced recovers from the file if it unexpectedly crashed and
restarted.

RestorableStateStorage caches process id of each thread in process_ids
field to find the tail thread entry in the file on compact(). The cache
Vec is required because traversing all process ids in the state hashmap
is inefficient. Also procfs can't be used because the thread of the tail
cell may be dead on compact() and fails to find the process id.

Alternative of compact() is changing the offset of the tail cell in the
state hashmap when a dead thread/process is removed from the hashmap.
However it is inefficient to have the offset of each cell as
`Rc<RefCell<usize>>` in terms of memory usage.

BUG=b:309571817
TEST=cros_run_unit_tests --board=$BOARD --packages resourced

Change-Id: Iee11d0afae51e9a489fd0f0596c964e3c30e7cbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5019640
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Youssef Esmat <youssefesmat@chromium.org>
Tested-by: Shin Kawamura <kawasin@google.com>
Commit-Queue: Shin Kawamura <kawasin@google.com>