biod: Refresh fingerprint templates on resume from hibernate
This change will monitor the power manager SuspendDone signal
and if the the deepest suspend state was to disk we will refresh the
fingerprint templates for the current user.

This happens by propagating the signal down to CrosFpSessionManager. We
only refresh the templates when the resume was from hibernate (to disk).

Instructions on manually hibernating can be found in the following
doc: go/hibernate-dogfood

BUG=b:301987542
TEST=Tested a hibernate on brya and confirmed biod reloads the
templates.

Change-Id: Idd88750020c13d23330ed8f24613c3bebf4a77ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5018647
Reviewed-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Tested-by: Brian Geffon <bgeffon@chromium.org>