shill: cellular: ignore duplicated entries in profile list update
The MBIM 1.0 profile management operations don't support IP type, and
this leads to us trying to connect with "vzwinternet" IPv4 instead of
the already existing "vzwinternet" IPv4v6 class 3 APN.

Due to this, we're notified of a profile list update right after we
connect the PDN, which shill takes as an indication to restart the
connection process with the new information.

This issue happens exclusively the first time we launch a connection
attempt. As soon as the new profile with the IPv4-only IP type is
stored, further updates in the list are ignored by shill.

We can workaround this issue simply by ignoring duplicated profile
entries in the profile list we receive from ModemManager. This
solution only makes sense because we're using MBIM 1.0 profile
management.

  2023-11-13T22:46:47.908601Z INFO ModemManager[2197]: <inf> [modem3] listed 3GPP profiles: 2 found
  2023-11-13T22:46:47.908614Z INFO ModemManager[2197]: <inf> [modem3] profile 0:
  2023-11-13T22:46:47.908630Z INFO ModemManager[2197]: <inf> [modem3]   profile id: 3
  2023-11-13T22:46:47.908639Z INFO ModemManager[2197]: <inf> [modem3]   apn: vzwinternet
  2023-11-13T22:46:47.908648Z INFO ModemManager[2197]: <inf> [modem3]   apn type: default
  2023-11-13T22:46:47.908692Z INFO ModemManager[2197]: <inf> [modem3] profile 1:
  2023-11-13T22:46:47.908706Z INFO ModemManager[2197]: <inf> [modem3]   profile id: 8
  2023-11-13T22:46:47.908716Z INFO ModemManager[2197]: <inf> [modem3]   apn: vzwinternet
  2023-11-13T22:46:47.908725Z INFO ModemManager[2197]: <inf> [modem3]   apn type: default
  2023-11-13T22:46:47.909383Z DEBUG shill[5106]: VERBOSE3 shill: [cellular_capability_3gpp.cc(2499)] /device/wwan0 OnProfilesListReply
  2023-11-13T22:46:47.909404Z DEBUG shill[5106]: VERBOSE3 shill: [cellular_capability_3gpp.cc(2217)] /device/wwan0 OnProfilesChanged
  2023-11-13T22:46:47.909431Z INFO shill[5106]: INFO shill: [cellular_capability_3gpp.cc(2270)] No update in stored profiles

BUG=b:307880765
TEST=Unit tests and manual testing with Uldren+VZW

Change-Id: I93e24451b824ebd6a7d19f1d5b334b2c2f36521d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5024754
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>