cryptohome: Serialize auth session work by username
This is a pretty significant rework of the way that the session manager
works under the covers. This replaces the current work queue system,
where there's a queue per session, with one where sessions are grouped
by username and there's a single queue per user.

The net effect of this is that you won't be able to have two sessions in
use at the same time if they for the same user. The plan is to
eventually take advantage of this to synchronize the AuthFactor and USS
bits between all sessions for a user, which is much easier if we can
assume that two different sessions are not fiddling with the same bits
at the same time.

In practice this shouldn't see a big impact on clients as all operations
were serialized onto the single mount thread anyway. It does reduce the
amount of things which can be interleaved, but overall impact on
responsivness should not be high.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib700e263b5e28b5754247c00150bae54dd5839a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5034285
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>