cryptohome: Use RunWhenAvailable for session creation
Strictly speaking this may seem unnecessary; when you create a session
you know it's not in use yet, so why make you jump through the hoops of
needing to schedule separately on it? But the problem is that we plan to
change meaning of "in use" to apply across all sessions for a user and
so it will be possible to create a new session that's not yet ready for
use, because the user already has another active in-use session.

This also cleans up the CreateAuthSession API to remove the error case,
which simplifies things. This change was avoided somewhat because it
required making a lot of test fixes, but this change had to update all
of those call sites anyway and so it was a natural thing to fix at the
same time.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ibc31be01f7a0f537876b1dbaf05624ad888fa023
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5031491
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>