shill: cellular: clarify Cellular.CustomAPNList API
The description of Cellular.CustomAPNList has a leftover from the
initial design that implies that the APN logic is in auto or manual mode
for each apn_type. The design was updated a while back to use manual
mode when ANY APN was in the list, irrespective of the apn_type.

BUG=None
TEST=None

Change-Id: Ic1fd1984a61a7b9a296c14401667927cb939b461
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5031253
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>