camera: Request CrosJpegAccelerator service only when it's available
In capture_unittests, the CrosJpegAccelerator service is disabled. It
makes the test overtime since JpegEncodeAccelerator waits 5 seconds
for the callback from the mojo channel. This CL let
JpegEncodeAccelerator query the service status first.
JpegEncodeAccelerator request the service only if the service is on.
It reduces the test time of UsingRealWebcam_TakePhoto from 7 secs to 2
secs.

BUG=b:228238413
TEST=Pass camera.CaptureUnittests camera.PlatformServiceSmoke on brya

Change-Id: Ie5442b06e676ceaaf7dbd9207e96b0353324fb8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5029591
Reviewed-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Sean Li <seannli@google.com>
Tested-by: Sean Li <seannli@google.com>