shill: handle better failures to start network validation
Network::StartPortalDetection() can fail to start if:
  - the network is disconnected,
  - there is no IP provisioned,
  - there is no DNS configured.

In all of these three cases, the Service should be considered as having
no connectivity for practical purposes if it still in an active state.

This patches ensures that the Service state is set to 'no-connectivity'
if its Network cannot start network validation with PortalDetector.

Until now the Device had optimistically assumed that the Service could
still be classified as online, but now that only network level issues
remains (for instance probe URL configuration errors were moved outside
of the network validation start path in crrev/c/4902868), it is safe to
consider the Service as no having no connectivity.

BUG=b:211000413
TEST=FEATURES=test emerge-$BOARD shill. Flashed trogdor, configured
ethernet with no DNS information (set IP provisioning to "manual", let
DNS settings to "automatic", and plugged ethernet dongle to an ethernet
port with no network), observed network validation failing to start and
the observed the Service transition to "no-connectivity".

Change-Id: I5d4ac456c63108df3f050a49b2a359026435aa4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5028989
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Cassie Wang <cassiewang@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>