net-base: parse link message by base::span<const uint8_t>
Originally, we use pointer arithmetic to parse the link attribute from
the rtnetlink packet. It's difficult to judge whether the code
doesn't touch the out-of-range memory.

This CL changes to use base::span<const uint8_t> to parse the packet.
With the hint of the valid data range, it's easier to judge whether the
code is memory-safe or not.

BUG=none
TEST=FEATURES=test emerge-brya net-base
TEST=/usr/libexec/fuzzers/rtnl_handler_fuzzer

Change-Id: Ia6c40bed2ca642d60e829a811e84bad8a4397cc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5029766
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>