printscanmgr: use executor to retrieve PPD files
Hook up the printscanmgr client to call into the executor to retrieve
PPD files. Now that two different tools use the executor's mojo remote,
change ownership of the remote to the D-Bus adaptor, which can then pass
unowned pointers to each of the tools that need it.

BUG=b:257070388
TEST=FEATURES=test emerge-nami printscanmgr

Change-Id: Ib29e4b8e27bd1dc39c11405b51e55d479c3ead30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5034383
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>