net-base: parse rule message by base::span<const uint8_t>
Originally, we use pointer arithmetic to parse the rule attribute
from the rtnetlink packet. It's difficult to judge whether the code
doesn't touch the out-of-range memory.

This CL changes to use base::span<const uint8_t> to parse the packet.
With the hint of the valid data range, it's easier to judge whether the
code is memory-safe or not.

BUG=none
TEST=FEATURES=test emerge-brya net-base
TEST=/usr/libexec/fuzzers/rtnl_handler_fuzzer

Change-Id: I72c3f42d6f8005050f2bd9a380f20b063ab53e8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5035758
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>