net-base: parse neighbor message by base::span<const uint8_t>
Originally, we use pointer arithmetic to parse the neighbor attribute
from the rtnetlink packet. It's difficult to judge whether the code
doesn't touch the out-of-range memory.

This CL changes to use base::span<const uint8_t> to parse the packet.
With the hint of the valid data range, it's easier to judge whether the
code is memory-safe or not.

BUG=none
TEST=FEATURES=test emerge-brya net-base
TEST=/usr/libexec/fuzzers/rtnl_handler_fuzzer

Change-Id: I542a4b5a32fcf89789c5139dc5808fa2dd631a75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5035877
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>