hiberman: Preload encrypted image data
On hibernate resume add a loop device on top of the hiberimage
LV and use it as backing device for the hiberimage. Start
reading encrypted hiberimage data before the use logs in to
populate the page cache with this data. This speeds up the
later loading of hibernate image after the user logged in,
since much of the data can be retrieved from the page cache
rather than loading it from disk.

BUG=b:311395327
TEST=hibernate on brya

Cq-Depend: chromium:5079343
Change-Id: Ibb98ef0d2ac6647719cdb89fa87275714f5e8ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5075780
Auto-Submit: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>