shill: portal: use same default HTTP URLs as Android
This patch changes the default HTTP URL used for portal detection to
http://connectivitycheck.gstatic.com/generate_204. This is the same URL
used by default in AOSP.

Now that network validation is not run on networks associated with
Services managed by policies, it is possible to align the portal URL
configuration with shill without risk of hitting on-prem transparent
proxies.

Using the same URLs as AOSP for portal redirection is expected to
increase the number of captive portal detected.

BUG=b:309175584
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: If7e783f8105ea1eb9e63947525e72ea75a10ebb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5052144
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>