modemfwd: Fixed flash modem when helper get_fw_info failed
modemfwd should not flash when the helper fails to get the version.

When the helper succeeds in fetching the version, but some partition
version is empty, modemfwd should not flash a partition where the
version has been obtained and the version is correct.

BUG=b:311595508
TEST=flash fires, and the helper output empty

Change-Id: I4158ba98772d7ed4b990392b03270e3f0a1eac1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5039673
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Madhav <madhavadas@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>