shill: portal: change probe result classification
This patch changes the classification of PortalDetector::Result such
that any partial failure is considered as a "no-connectivity" result.

Chrome is currently interpreting "portal-suspected" results by inviting
the user to sign into a supposed captive portal, However, this behavior
has multiple false positive scenarios:
  - If the network has no captive portal but low bandwidth and the
    probes can genuinely fail by timing out while other probes succeed,
    a portal signin notification should not be shown.
  - If the network has some form of transparent proxy or firewall
    rewriting HTTP 204 responses as 200 responses with no content, the
    result should be considered a success. A portal signin notification
    will most likely lead to the "white page portal" symptom.

The only appropriate scenario where showing the signin notification for
a response that is not a redirection with a Location header is when the
HTTP probe receives a 200 response with content in the response body. A
follow-up patch will add logic to detect such responses and classify
the as "portal-suspected". Some unit tests are temporarily disabled
until a new "portal-suspected" state based on Content-Length checks is
introduced.

BUG=b:299851050
BUG=b:309175584
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I6cb292c727ba529afe41589eba8c37bf784eba95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5052148
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>