vm_tools: concierge: Delay checking VMMMS feature
Delay checking the VMMMS feature until the last moment, when ARCVM is
about to be launched. This is necessary because chrome://flags are not
immediately available when concierge starts. Delaying the check
requires that GetVmMemoryManagementKillsConnection be made async so that
the response can be delayed until the initialization check. Note that
this means the call will take longer or may even never finish. However,
resourced (the only client) does the call via an async await, so it can
already handle this delay.

BUG=b:311935336
TEST=Enable VMMMS in chrome://flags, re-login, observe logs

Change-Id: I43fbf9f3569c6fb5b6720ac01e9921b3e95439b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5039404
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>
Auto-Submit: David Stevens <stevensd@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>