platform2: Remove old references to "schedtune"
Everything related to "schedtune" was just there to support an
interface that was present in downstream kernels v4.19 and below. This
code was only ever enabled on ARM kernels.

All ARM boards have now migrated to kernel v5.4 and newer and the old
code is no longer needed. It can be confirmed that grepping for
"schedtune" on kernels v5.4 and newer returns no hits. Delete the old
code.

BUG=b:271420003
TEST=CQ

Change-Id: Ifed5c7340949273dc04f4b89ba0b266a368a4ce3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5050688
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>