shill: wifi: fix Shill crash due to HotspotDevice use after free
HotspotDevice's lifetime is controlled by WiFiProvider. The instance
could be destroyed before the pending callbacks are called. Per
//docs/callback.md, when binding with base::Unretained, there must be
some external guarantee that the class instance will always be live
when running the callback, which is not true in HotspotDevice's case.
If there is no external lifetime guarantee, base::WeakPtr is suggested.
If this is destroyed before the posted callback runs, the callback will
simply become a no-op when run, rather than being a use-after-free bug
on the destroyed instance.

TEST=FEATURES=test emerge-brya shill
BUG=b:312430584

Change-Id: I0d58213fcffd95af707a5d4870459f04ad5ed210
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5051634
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>