shill: portal: redefine HttpRequest Error enum
This patch refactors the HttpRequest::Result enum such that:
  - it is an enum class
  - unused values are removed
  - the names match better the underlying error
  - the success cases are removed

This allows this enum to properly express an error case for the
HttpRequest class. PortalDetector was already treating this enum as an
error value.

This refactoring will allow replacing the PortalDetector::Phase and
PortalDetector::Status values with better defined enums specific to the
HTTPS probe and HTTP probe.

BUG=b:309175584
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I68540ef74d496d3aaaf262c162132e77826a9d49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5052146
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>