libbrillo: Make a crash-report if variant unwrapping implicitly happens.
The part is planned to be removed, while it is difficult to change
step-by-step unfortunately, because it is used in generated code.
This CL adds crash report (without changing the production behavior),
so we can make sure there are no actual uses in production.

BUG=b:289932268
TEST=Tryjob

Change-Id: I37dd4058119841988561687b377095b779e23318
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5051207
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>