power: Store the system default mem_sleep value
Keep track of the kernel's initial setting for `/sys/power/mem_sleep`.
Store the value only on first powerd on boot as the contents will change
based on policy when powerd suspends the system.

The reven board runs on a variety of hardware and the kernel's default
setting for mem_sleep will vary from machine to machine.
Storing this value will allow future powerd policies the ability to
dynamically defer to the kernel's choice for the device.

BUG=b:275614676
TEST=`cros_workon_make --board=reven power_manager --test`
TEST=`USE="asan fuzzer" cros build-packages --board=reven
--skip-chroot-upgrade power_manager`

Change-Id: Ieebb581b8bf4f754b60482a389fc5f810095ce0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5055445
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>